Add other lambdas.
